server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
monai
nibabel
numpy
Pillow
matplotlib
scipy
scikit-image
trimesh

